+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Official support for Python 3.7.
-- Add optional `hostname` parameter to `ctds.connect`.
+- Add optional `hostname` parameter to `ctds.connect()`.
 
 ### Fixed
 - Retry on SQL Server unittest database setup failures due to race conditions
@@ -19,6 +19,8 @@ builds (FTP no longer works due to firewall issues.)
 - Fix OS X Travis CI build.
 - Improve SQL Server -> DB API 2.0 error mappings.
 - Remove Python 2.6 from Appveyor CI due to lack of support.
+- Fix `ctds.Cursor.executemany()` truncating data for variable width types.
+https://github.com/zillow/ctds/issues/25.
 
 ## [1.7.0] - 2018-01-24
 ### Added